requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-filter==23.2
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
